refactor: User관련 Json 파일들을 합쳤습니다. Token과 AccessToken를 token.dart 파일로, User와 ModifyUser, 그리고 UserEmail을 user.dart로 한 파일로 합쳤습니다.
