--- lib/app/data/model/json_models/user/access_token_model.dart
@@ -1,17 +1,0 @@
-import 'package:json_annotation/json_annotation.dart';
-
-part 'access_token_model.g.dart';
-
-@JsonSerializable()
-class AccessToken {
-  String access_token;
-
-  AccessToken({
-    required this.access_token,
-  });
-
-  factory AccessToken.fromJson(Map<String, dynamic> json) =>
-      _$AccessTokenFromJson(json);
-
-  Map<String, dynamic> toJson() => _$AccessTokenToJson(this);
-}
