+++ lib/app/data/model/json_models/user/user_model.dart
@@ -20,3 +20,37 @@ class User {
 
   Map<String, dynamic> toJson() => _$UserToJson(this);
 }
+
+@JsonSerializable()
+class UserEmail {
+  String email;
+
+  UserEmail({
+    required this.email,
+  });
+
+  factory UserEmail.fromJson(Map<String, dynamic> json) =>
+      _$UserEmailFromJson(json);
+
+  Map<String, dynamic> toJson() => _$UserEmailToJson(this);
+}
+
+@JsonSerializable()
+class ModifyUser {
+  String email;
+  String nickname;
+  String tel;
+  String image;
+
+  ModifyUser({
+    required this.email,
+    required this.nickname,
+    required this.tel,
+    required this.image,
+  });
+
+  factory ModifyUser.fromJson(Map<String, dynamic> json) =>
+      _$ModifyUserFromJson(json);
+
+  Map<String, dynamic> toJson() => _$ModifyUserToJson(this);
+}
